[Topics] Fix ComsumeChannel coordination. Fixes #3
The tests were not flacky, the code was. We had a few issues because I
did not pay attention on how we consume the messages and how we can
ensure that we never drop messages:

\# Bugs

1. We were consuming messages withouth workers: dropped messages
2. We were not ensuring that the Create method would return once we
   started consuming.

To fix the above we had to introduce to changes:

1. If we have no workers, we do not consume messages and let the channel
   buffer.
2. To be able to coordinate the Consume method and the Create method we
   added a TaskCompletionSource that will only be set when we have data
   to consume.
3. If we start with an empty channel, send a Ack message to let the
   Consume method be ready and the TCS be set.

We have added a struct that wraps the messages default types, this will
give use more flexibility later to allow message retries.