[API] Be more extrict with the type of messages we allow. (#11)
It is very tempting to allow user to use reference types for the message queues, unfortunately that will mean that we could be paying a GC.Collect penalty if a low of small fast objects are used. Due to the nature of what we want to achieve is best to for the users to use value types that will be placed in the stack.