[API] Add the ability to close channels and wait for all events to be flushed (#21)
We add a new API that will allow developers to close a channel or all the
channels and wait until all the events to be flushed. This API will
always flush messages. We can later add an API that does not flushes the
messages by using the cancellation token rather than closing the
channel.