[API] Provide an error handling API. Fixes #31
In order to be able to handle errors/exception from workers we are going
to provide an API that allows to register an ErrorWorker per topic.

This API allows a worker that is registed to a topic to throw an
exception, this exception later will be wrapped in an event passed to
the error worker. The event will provide the following data:

1. Message: The message that generated the exception.
2. The exception that was thrown.