[API] Make IWorkers and IErrorWorkers implement IDisposable. (#44)
This change makes the Hub and a few other internal methods be
IDisposable too. Adding them as IDisposable allows use to later have
IDisposable referenes in workers which is very useful in certain
scenarions.

This change has also made the closing of the topics more realiable since
we now do wait after we close a channel. If we do not wait for the task
that consumes the channel, the task cannot be disposed. This approacj is
a lot nicer.

Fixes #10