[API] Allow workers to be CPU intensive. (#51)
Rather than just dealing with IO bound workers we are opening the door
to CPU intensive workers.

The new API allows the developer to let the hub know if the worker
callback should be executed in a background task or not. This allow
users to make the decision to use or no the main thread used by the hub.

There is not need to move the Consuming methods out of the main thread
because those are IO bound by definition (readin gfrom a channel is IO
bound). The only code that should be executed in a diff thread is that
of the consumer.